tools/book_import/pyproject.toml
[project]
name = "book-import"
version = "0.1.0"
description = "Import public-domain books into liberalismo.info markdown schema"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "click>=8.1.8",
  "ebooklib>=0.18",
  "lxml>=5.3.0",
  "pypdf>=5.1.0",
  "rapidfuzz>=3.10.1",
  "requests>=2.32.3",
]

[project.optional-dependencies]
ocr = [
  "pillow>=11.0.0",
  "pytesseract>=0.3.13",
  "pymupdf>=1.24.12",
]

[project.scripts]
book-import = "book_import.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/book_import"]
